refactor MovieDetailPage and create MovieDetail component
