+++ project-movie-app/src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './index.css'
 import { App } from './App.jsx'
-import { MovieDetailPage } from './CardDetailPage/MovieDetailPage.jsx';
+import { MovieDetailPage } from './MovieDetailPage/MovieDetailPage.jsx';
 
 const root = document.getElementById("root");
